ensure globs work with absolute patterns
If a glob starts with `./`, `../` or normal strings like `*.ts`, then we
have to prepend the base path. However, if a glob is an absolute pattern
such as `/Users/tailwind/dev/project/**/*.html`, then we don't have to
prepend the base path.

This is useful for when we are working with `@content`, and we can
prepend the location of the current file to the glob. That way it will
already be scoped to the correct location.