use a macrotask instead of microtask
If you happen to have globs that match the same base path, it could
result in a file that is picked up by multiple watchers. We used a
microtask to collect all of them but that's not enough.

Instead, at least on macOS, a macrotask is enough (implemented as a
`setTimeout(cb, 0)`). Even if we trigger a build multiple times it
should result in a no-op the second time so it's not the end of the
world.

This is a bit nicer, because now you only see a single output log in
instead of `n`.