improve themeOptions usage
Destructuring works, but it's creating an unnecessary array. We are
currently interested in whether the `inline` and/or `reference` values
exist in the `themeOptions` list.

We already know that the very first item is going to be `@theme`, so we
can skip over that by providing the `fromIndex` when using
`.includes(value, fromIndex)`