add `waitForOutputFileChange`
Instead of relying on CLI output, let's rely on changes made to the File
System itself.

The API looks like:

```
await waitForOutputFileChange(path, cb => {
  // Do something
})
```

It will setup a watcher for the given `path`, once the watcher is ready,
we will invoke the callback.

Once the file on disk changes, we can resolve the wrapping promise.

We also have this helper in the v3 codebase, a rename might be good,

Let's see if this fixes flakey tests on CI or not.