remove top-level `scan_files`
The only place we used this in was in the UI tests. This is also the
only spot where we used the public `IO` and `Parsing` enums.

Note: you'll notice that I temporary used the `scanDir()` function in
the UI test which is a bit ugly, but because we currently don't pass in
any `sources` or `base` options, it's almost a no-op. But it does expose
the `scanFiles` function we use for incremental rebuilds.

This will be cleaned up in the next few commits