handle `@apply` once
This now collects the custom `@utility` rules if they contain `@apply`.
Once all the `@utility` utilities are registered in the system, then we
will update (mutate) the `node` to replace all the `@apply` rules with
its substitute.

This means that by the time you are using the `@utility`, the `@apply`
is already handled and we don't pay any penalties that require handling
the `@apply`.

We wait for all `@utility` utilities to be registered, so that we can
use one `@utility` via `@apply` in another (out of order).

We will also verify that we don't run into circular dependencies
upfront.