replace `@utility` after we handled `@apply`
We can use `@apply` inside of `@utility` calls. Before this we collected
the `@utility` before replacing it with `[]`.

Instead, we don't mutate the AST yet, handle `@apply` once, last we
remove the `@utility` at-rules. Because we know that `@utility` has to
be top-level, we can also improve the performance of the `walk` by never
traversing into nested nodes.