Add `matchUtilities` and `addUtilities` APIs (#14114)
This PR introduces support for the v3-like `addUtilities` and
`matchUtilities` APIs in v4. We anticipate designing a new API that
feels more native to the way v4 works before shipping v4.0 stable, but
we're continuing to support these APIs for backwards compatibility.

We've tried to make the behavior as identical as possible, but because
of fundamental differences between the v3 and v4 engines there are a few
things that work differently:

## Only simple single-class selectors are supported

In v3 you could pass a complex CSS selector to `addUtilities` and we
would generate a utility for every class in the selector. In v4 we only
allow you to use a simple, single-class selector.

You should use nesting if you need a more complex selector, or need to
include at-rules like `@media` or `@supports`.

```js
// v3
function ({ addUtilities }) {
  addUtilities({
    '.scrollbar-none::-webkit-scrollbar': {
      display: 'none',
    },
  })
}

// v4
function ({ addUtilities }) {
  addUtilities({
    '.scrollbar-none': {
      '&::-webkit-scrollbar': {
        display: 'none',
      },
    },
  })
}
```

If you were adding custom utilities that included two classes and were
depending on both of those classes behaving like utilities (they could
each be used with variants), those custom utilities will need to be
rewritten as two separate utilities that each use nesting:

```js
// v3
function ({ addUtilities }) {
  addUtilities({
    '.section > .row': {
      color: 'red',
    },
  })
}

// v4
function ({ addUtilities }) {
  addUtilities({
    '.section': {
      '& > .row': {
        color: 'red',
      },
    },

    '.row': {
      'section > &': {
        color: 'red',
      },
    },
  })
}
```

We may introduce support for this in the future if this limitation turns
out to be a huge pain in the ass, but crossing our fingers that people
were mostly doing simple stuff here.

## Opacity modifiers support bare values

To be consistent with how built-in utilities work in v4, custom
utilities that specify `type: "color"` now get "bare value" support for
opacity modifiers. This means that a utility like `foo-black/33` will
work out of the box without having to either add `--opacity-33` to your
theme nor would you need to add it to the `modifiers` option.

## The `preferOnConflict` type option is gone

In v3 we introduced an internal API called `preferOnConflict` for types.
This was used as a way to disambiguate between two utilities with the
same "root" but which produced different properties which used the same
CSS data types. This was only applicable to arbitrary values and was
only used for disambiguating between `background-position` and
`background-size`.

In v4, both of these properties are handled by a single plugin meaning
this feature is no longer necessary. No one should've really been using
this option anyway as it was never documented so we're dropping the
feature.

## The options `respectPrefix` and `respectImportant` are not yet
supported

Neither the `prefix` nor `important` features exist in any form in v4 at
this time. Therefore, they are not currently supported by this PR. We
will look into supporting them if/when those features return.

## The `theme(…)` function is not currently supported

Custom utilities defined using `matchUtilities` often use the `theme(…)`
function to define their default values, but we haven't implemented
support for `theme(…)` yet in v4.

This means that as of this PR, default values for custom utilities must
be hardcoded:

```js
function ({ matchUtilities }) {
  matchUtilities({
    'tab': (value) => {
      return {
        'tab-size': value,
      }
    },
  }, {
    values: {
      2: '2',
      4: '4',
      8: '8',
    },
  })
}
```

Getting `theme(…)` working is a big project so we're going to tackle it
in a separate PR.

---------

Co-authored-by: Adam Wathan <4323180+adamwathan@users.noreply.github.com>
Co-authored-by: Robin Malfait <malfait.robin@gmail.com>