Retry UI test assertions (#14165)
We noticed that some of the tests depending on the `hover:` variant were
flaky. After some investigation, we found that injected elements had the
`:hover` state without us explicitly hovering over the element.

To avoid this, we now set up an explicit placeholder element to move the
mouse to before running the tests.